Updated Util Files

+++ include/util.c
@@ -9,9 +9,9 @@ void memory_copy(char *source, char *dest, int nbytes)
     }
 }
 
-void memory_set(uint8 *dest, uint8 val, uint32 len)
+void memory_set(uint32 *dest, uint32 val, uint32 len)
 {
-    uint8 *temp = (uint8 *)dest;
+    uint32 *temp = (uint32 *)dest;
     for ( ; len != 0; len--)
     {
       *temp++ = val;
